+++ apps/web/components/sections/home/community.tsx
@@ -1,10 +1,11 @@
 'use client';
 
-import { motion } from 'framer-motion';
+import { motion, useReducedMotion } from 'framer-motion';
 import { Rocket, Shield, TrendingUp, Users } from 'lucide-react';
 import { BenefitItem } from '~/components/shared/benefits-items';
 import { CTAForm } from '~/components/shared/cta-form';
 import { Testimonial } from '~/components/shared/testimonial-card';
+import { useState } from 'react';
 
 // Constants
 const benefits: Benefit[] = [
@@ -40,6 +41,19 @@ const testimonialData: TestimonialData = {
 
 // Component
 export function Community() {
+  const prefersReducedMotion = useReducedMotion();
+  const [formStatus, setFormStatus] = useState<FormStatus | null>(null);
+
+  const handleFormSubmission = async (data: FormData) => {
+    setFormStatus(null); // Reset the status before starting the submission
+    try {
+      await submitForm(data);
+      setFormStatus({ type: 'success', message: 'Form submitted successfully!' });
+    } catch (error) {
+      setFormStatus({ type: 'error', message: 'Failed to submit the form. Please try again.' });
+    }
+  };
+
   return (
     <section className="py-20">
       <div className="container mx-auto px-4">
@@ -48,7 +62,9 @@ export function Community() {
           initial={{ opacity: 0, y: 20 }}
           whileInView={{ opacity: 1, y: 0 }}
           viewport={{ once: true }}
-          transition={{ duration: 0.6 }}
+          transition={{
+            duration: prefersReducedMotion ? 0 : 0.6, // Respects reduced motion preferences
+          }}
           className="text-center mb-20 max-w-3xl mx-auto"
         >
           <h2 className="text-4xl font-bold text-gray-900 mb-6">
@@ -86,10 +102,35 @@ export function Community() {
         </div>
 
         {/* CTA Form */}
-        <CTAForm onSubmit={(data) => console.log(data)} />
+        <CTAForm onSubmit={handleFormSubmission} />
+
+        {/* Form Status Message */}
+        {formStatus && (
+          <div
+            className={`mt-4 p-4 rounded-lg ${
+              formStatus.type === 'success' ? 'bg-green-100 text-green-800' : 'bg-red-100 text-red-800'
+            }`}
+          >
+            {formStatus.message}
+          </div>
+        )}
       </div>
     </section>
   );
+}
+
+// Mock form submission function
+const submitForm = async (data: FormData): Promise<void> => {
+  // Simulate a delay
+  return new Promise((resolve, reject) => {
+    setTimeout(() => {
+      if (Math.random() > 0.5) {
+        resolve();
+      } else {
+        reject(new Error('Submission failed'));
+      }
+    }, 1000);
+  });
 };
 
 // Interfaces
@@ -104,4 +145,13 @@ interface TestimonialData {
   author: string;
   role: string;
   imageUrl: string;
+}
+
+interface FormData {
+  [key: string]: string | number; // Replace with specific form fields if known
+}
+
+interface FormStatus {
+  type: 'success' | 'error';
+  message: string;
 }
