+++ apps/web/components/sections/home/highlighted-projects.tsx
@@ -11,154 +11,7 @@ import {
 import { CTAButtons } from "~/components/shared/cta-buttons";
 import { ProjectCard } from "~/components/shared/project-card";
 import { SectionCaption } from "~/components/shared/section-caption";
-
-// Constants
-const projects: Project[] = [
-  {
-    id: "healthy-kids-id",
-    image: "/images/kids.webp",
-    category: "Child Welfare",
-    title: "Healthy Kids Workshop",
-    description:
-      "Provide nourishing meals and support to children at risk of malnutrition in Costa Rica. Together, we can ensure a brighter future for every child.",
-    currentAmount: 22800,
-    targetAmount: 25000,
-    investors: 18,
-    minInvestment: 5,
-    percentageComplete: 90,
-    tags: [
-      { id: "ngo-tag-id", text: "NGO" },
-      { id: "nutrition-tag-id", text: "NUTRITION" },
-      { id: "children-tag-id", text: "CHILDREN" },
-    ],
-  },
-  {
-    id: "forest-restoration-id",
-    image: "/images/bosques.webp",
-    category: "Environmental Protection",
-    title: "Forest Restoration Initiative",
-    description:
-      "Restore and reforest areas devastated by uncontrolled deforestation. Your support helps rebuild ecosystems and fight climate change.",
-    currentAmount: 54000,
-    targetAmount: 60000,
-    investors: 35,
-    minInvestment: 10,
-    percentageComplete: 90,
-    tags: [
-      { id: "environment-tag-id", text: "ENVIRONMENT" },
-      { id: "ecological-tag-id", text: "ECOLOGICAL" },
-      { id: "sustainable-tag-id", text: "SUSTAINABLE" },
-    ],
-  },
-  {
-    id: "rural-animal-shelter-id",
-    image: "/images/dogs.webp",
-    category: "Animal Welfare",
-    title: "Rural Animal Shelter",
-    description:
-      "Provide care and shelter to homeless animals in rural communities. Help us create safe havens for animals in need.",
-    currentAmount: 15500,
-    targetAmount: 20000,
-    investors: 22,
-    minInvestment: 8,
-    percentageComplete: 77,
-    tags: [
-      { id: "animals-tag-id", text: "ANIMALS" },
-      { id: "care-tag-id", text: "CARE" },
-      { id: "community-tag-id", text: "COMMUNITY" },
-    ],
-  },
-  {
-    id: "disaster-aid-id",
-    image: "/images/disaster-aid.webp",
-    category: "Disaster Relief",
-    title: "Natural Disasters Human Aid",
-    description:
-      "Provide critical support to communities affected by natural disasters. From emergency supplies to long-term rebuilding efforts, join us in bringing hope and recovery to those in need.",
-    currentAmount: 30000,
-    targetAmount: 50000,
-    investors: 28,
-    minInvestment: 20,
-    percentageComplete: 60,
-    tags: [
-      { id: "humanitarian-tag-id", text: "HUMANITARIAN" },
-      { id: "disaster-tag-id", text: "DISASTER RELIEF" },
-      { id: "community-tag-id", text: "COMMUNITY SUPPORT" },
-    ],
-  },
-  {
-    id: "indigenous-crafts-id",
-    image: "/images/artesania.webp",
-    category: "Culture and Arts",
-    title: "Preserving Indigenous Crafts",
-    description:
-      "Support the preservation of indigenous craftsmanship in Costa Rica. Your contributions protect traditional techniques and cultural heritage.",
-    currentAmount: 34000,
-    targetAmount: 50000,
-    investors: 29,
-    minInvestment: 15,
-    percentageComplete: 68,
-    tags: [
-      { id: "culture-tag-id", text: "CULTURE" },
-      { id: "indigenous-tag-id", text: "INDIGENOUS" },
-      { id: "art-tag-id", text: "ART" },
-    ],
-  },
-  {
-    id: "water-for-rural-communities-id",
-    image: "/images/water.webp",
-    category: "Access to Clean Water",
-    title: "Water for Rural Communities",
-    description:
-      "Provide access to safe drinking water in underserved rural areas. Help us install water purification systems to improve health and livelihoods.",
-    currentAmount: 18500,
-    targetAmount: 25000,
-    investors: 20,
-    minInvestment: 12,
-    percentageComplete: 74,
-    tags: [
-      { id: "water-tag-id", text: "WATER" },
-      { id: "health-tag-id", text: "HEALTH" },
-      { id: "community-tag-id", text: "COMMUNITY" },
-    ],
-  },
-  {
-    id: "empowering-education-id",
-    image: "/images/education.webp",
-    category: "Education",
-    title: "Empowering Education",
-    description:
-      "Support education programs for children in low-income areas. Together, we can bridge the education gap and create opportunities for the next generation.",
-    currentAmount: 40000,
-    targetAmount: 55000,
-    investors: 40,
-    minInvestment: 10,
-    percentageComplete: 73,
-    tags: [
-      { id: "education-tag-id", text: "EDUCATION" },
-      { id: "children-tag-id", text: "CHILDREN" },
-      { id: "future-tag-id", text: "FUTURE" },
-    ],
-  },
-  {
-    id: "mobile-clinics-id",
-    image: "/images/healthcare.webp",
-    category: "Healthcare",
-    title: "Mobile Clinics",
-    description:
-      "Bring essential healthcare services to remote areas through mobile clinics. Your support helps save lives and build healthier communities.",
-    currentAmount: 32000,
-    targetAmount: 45000,
-    investors: 30,
-    minInvestment: 20,
-    percentageComplete: 71,
-    tags: [
-      { id: "healthcare-tag-id", text: "HEALTHCARE" },
-      { id: "community-tag-id", text: "COMMUNITY" },
-      { id: "impact-tag-id", text: "IMPACT" },
-    ],
-  },
-];
+import { projects } from "~/lib/mock-data/mock-projects";
 
 // Component
 export function HighlightedProjects() {
@@ -183,19 +36,28 @@ export function HighlightedProjects() {
               }),
             ]}
             className="w-full"
+            aria-label="Featured Projects"
           >
             <CarouselContent className="-ml-2 md:-ml-4">
               {projects.map((project) => (
                 <CarouselItem
                   key={project.id}
                   className="pl-2 md:pl-4 basis-full sm:basis-1/2 lg:basis-1/3 xl:basis-1/4"
+                  role="group"
+                  aria-roledescription="slide"
                 >
                   <ProjectCard {...project} />
                 </CarouselItem>
               ))}
             </CarouselContent>
-            <CarouselPrevious className="hidden sm:flex" />
-            <CarouselNext className="hidden sm:flex" />
+            <CarouselPrevious
+              className="hidden sm:flex"
+              aria-label="Previous slide"
+            />
+            <CarouselNext
+              className="hidden sm:flex"
+              aria-label="Next slide"
+            />
           </Carousel>
         </div>
 
@@ -209,23 +71,3 @@ export function HighlightedProjects() {
     </section>
   );
 };
-
-// Interfaces
-interface Tag {
-  id: string;
-  text: string;
-}
-
-interface Project {
-  id: string;
-  image: string;
-  category: string;
-  title: string;
-  description: string;
-  currentAmount: number;
-  targetAmount: number;
-  investors: number;
-  minInvestment: number;
-  percentageComplete: number;
-  tags: Tag[];
-}
